apps/chatstreet-backend/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.0
requests==2.28.1
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.13
Werkzeug==2.2.3
wget==3.2
